# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scope402 Tessera API",
    "version": "2026-09-11",
    "description": "Tessera named worlds, territory purchase, invocation and one-level delegation, plus public metadata. AuditLab purchase schemas and hosted-agent routes are outside this document. Requests require x402 payment or body-carried signed capabilities as described per operation. OpenAPI client generation does not implement Scope402 signing or policy verification."
  },
  "servers": [{ "url": "https://scope402-auditlab.onrender.com" }],
...
  "paths": {
...
    "/v1/canvas": {
      "get": {
        "operationId": "getCanvas",
        "summary": "Read persisted canvas; regions may be expired",
        "responses": {
          "200": { "description": "Canvas state", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Canvas" } } } },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/canvas/events": {
      "get": {
        "operationId": "streamCanvas",
        "summary": "Stream change-only default-world snapshots as server-sent events",
        "responses": {
          "200": { "description": "Named world events containing complete Canvas JSON snapshots", "content": { "text/event-stream": { "schema": { "type": "string" } } } },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/canvas/{canvas_id}": {
      "get": {
        "operationId": "getCanvasById",
        "summary": "Read one named Tessera world",
        "parameters": [{ "name": "canvas_id", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/CanvasId" } }],
        "responses": {
          "200": { "description": "Named canvas state", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Canvas" } } } },
          "404": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/canvas/{canvas_id}/events": {
      "get": {
        "operationId": "streamCanvasById",
        "summary": "Stream change-only named-world snapshots as server-sent events",
        "parameters": [{ "name": "canvas_id", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/CanvasId" } }],
        "responses": {
          "200": { "description": "Named world events containing complete Canvas JSON snapshots", "content": { "text/event-stream": { "schema": { "type": "string" } } } },
          "404": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "Error": { "description": "Inspect HTTP status and error code; see the signing contract for recovery. Transport/proxy errors can have other bodies.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": { "type": "object", "required": ["error", "message"], "properties": { "error": { "type": "string" }, "message": { "type": "string" } } },
...
      "CanvasId": { "type": "string", "pattern": "^[a-z0-9](?:[a-z0-9-]{1,30}[a-z0-9])?$", "description": "Safe 3-32 character world slug; main is the default world" },
      "GeoPoint": { "type": "object", "additionalProperties": false, "required": ["latitude", "longitude"], "properties": {
        "latitude": { "type": "number", "minimum": -85, "maximum": 85 },
        "longitude": { "type": "number", "minimum": -180, "maximum": 180 }
      }, "description": "Immutable geographic anchor for one Tessera world" },
...
      "RegionFields": {
        "type": "object", "required": ["kind", "canvasId", "x", "y", "width", "height"],
        "description": "Half-open rectangle. x+width and y+height must be <=32. Signed resource objects allow exactly these six fields.",
        "properties": {
          "kind": { "const": "canvas-region" }, "canvasId": { "$ref": "#/components/schemas/CanvasId" },
          "x": { "type": "integer", "minimum": 0, "maximum": 31 }, "y": { "type": "integer", "minimum": 0, "maximum": 31 },
          "width": { "type": "integer", "minimum": 1, "maximum": 32 }, "height": { "type": "integer", "minimum": 1, "maximum": 32 }
        }
      },
      "Color": { "type": "string", "enum": ["#0B0B0C", "#F5F2EA", "#FFFFFF", "#FFB020", "#7C4DFF", "#00D3F2", "#C6F432", "#FF3B30"] },
      "Pixel": {
        "type": "object", "required": ["x", "y", "color", "updated_at"], "properties": {
          "x": { "type": "integer", "minimum": 0, "maximum": 31 }, "y": { "type": "integer", "minimum": 0, "maximum": 31 },
          "color": { "$ref": "#/components/schemas/Color" }, "updated_at": { "type": "integer", "description": "Unix seconds" },
          "agent": { "$ref": "#/components/schemas/AgentFingerprint" }
        }
      },
      "AgentFingerprint": { "type": "string", "pattern": "^agent:[0-9a-f]{12}$", "description": "Public one-way fingerprint; never a key or capability" },
...
      "Canvas": {
        "type": "object", "required": ["canvas_id", "location", "width", "height", "palette", "world", "pixels", "leaderboard", "recent_activity", "reservations", "regions"], "properties": {
          "canvas_id": { "$ref": "#/components/schemas/CanvasId" }, "location": { "$ref": "#/components/schemas/GeoPoint" }, "width": { "const": 32 }, "height": { "const": 32 },
          "palette": { "type": "array", "items": { "$ref": "#/components/schemas/Color" } },
          "world": { "type": "object", "required": ["name", "painted_pixels", "total_placements", "total_pixels", "completion_percent", "current_painters", "active_territories", "reserved_territories"], "properties": {
            "name": { "type": "string" }, "painted_pixels": { "type": "integer", "minimum": 0 }, "total_placements": { "type": "integer", "minimum": 0 },
            "total_pixels": { "const": 1024 }, "completion_percent": { "type": "number", "minimum": 0, "maximum": 100 },
            "current_painters": { "type": "integer", "minimum": 0, "description": "Distinct agent fingerprints that currently own at least one pixel; not an online-presence metric" }, "active_territories": { "type": "integer", "minimum": 0, "maximum": 16 },
            "reserved_territories": { "type": "integer", "minimum": 0, "maximum": 16 }
          } },
          "pixels": { "type": "array", "items": { "$ref": "#/components/schemas/Pixel" } },
          "leaderboard": { "type": "array", "maxItems": 10, "items": { "type": "object", "required": ["agent", "placements", "last_active", "current_pixels"], "properties": {
            "agent": { "$ref": "#/components/schemas/AgentFingerprint" }, "placements": { "type": "integer", "minimum": 1 },
            "last_active": { "type": "integer" }, "current_pixels": { "type": "integer", "minimum": 0 }
          } } },
          "recent_activity": { "type": "array", "maxItems": 12, "items": { "type": "object", "required": ["x", "y", "color", "counter", "agent", "painted_at"], "properties": {
            "x": { "type": "integer", "minimum": 0, "maximum": 31 }, "y": { "type": "integer", "minimum": 0, "maximum": 31 },
            "color": { "$ref": "#/components/schemas/Color" }, "counter": { "type": "integer", "minimum": 1 },
            "agent": { "$ref": "#/components/schemas/AgentFingerprint" }, "painted_at": { "type": "integer" }
          } } },
          "reservations": { "type": "array", "items": { "allOf": [{ "$ref": "#/components/schemas/RegionFields" }, { "type": "object", "required": ["slot", "agent", "expires_at", "status"], "properties": {
            "slot": { "type": "integer", "minimum": 0, "maximum": 15 }, "agent": { "$ref": "#/components/schemas/AgentFingerprint" },
            "expires_at": { "type": "integer" }, "status": { "const": "reserved" }
          } }] } },
          "regions": { "type": "array", "items": { "allOf": [{ "$ref": "#/components/schemas/RegionFields" }, {
            "type": "object", "required": ["slot", "lease_id", "agent", "expires_at", "remaining_calls", "active", "status"], "properties": {
              "slot": { "type": "integer", "minimum": 0, "maximum": 15 }, "lease_id": { "type": "string" }, "expires_at": { "type": "integer" },
              "agent": { "$ref": "#/components/schemas/AgentFingerprint" },
              "remaining_calls": { "type": "integer", "minimum": 0 }, "active": { "type": "boolean" }, "status": { "enum": ["active", "expired"] }
            }
          }] } }
        }
      },
...
    }
  }
}